+++ subprojects/core/src/main/java/org/gradle/api/internal/changedetection/state/FileCollectionSnapshotBuilder.java
@@ -16,41 +16,6 @@
 
 package org.gradle.api.internal.changedetection.state;
 
-import com.google.common.collect.Maps;
-import org.gradle.api.internal.cache.StringInterner;
-
-import java.util.Map;
-
-public class FileCollectionSnapshotBuilder {
-    private final Map<String, NormalizedFileSnapshot> snapshots = Maps.newLinkedHashMap();
-    private final PathNormalizationStrategy pathNormalizationStrategy;
-    private final StringInterner stringInterner;
-    private final TaskFilePropertyCompareStrategy compareStrategy;
-
-    public FileCollectionSnapshotBuilder(TaskFilePropertyCompareStrategy compareStrategy, PathNormalizationStrategy pathNormalizationStrategy, StringInterner stringInterner) {
-        this.pathNormalizationStrategy = pathNormalizationStrategy;
-        this.stringInterner = stringInterner;
-        this.compareStrategy = compareStrategy;
-    }
-
-    public void collectFileSnapshot(FileSnapshot fileSnapshot) {
-        String absolutePath = fileSnapshot.getPath();
-        if (!snapshots.containsKey(absolutePath)) {
-            NormalizedFileSnapshot normalizedSnapshot = pathNormalizationStrategy.getNormalizedSnapshot(fileSnapshot, stringInterner);
-            collectNormalizedFileSnapshot(absolutePath, normalizedSnapshot);
-        }
-    }
-
-    public void collectNormalizedFileSnapshot(String absolutePath, NormalizedFileSnapshot normalizedSnapshot) {
-        if (normalizedSnapshot != null && !snapshots.containsKey(absolutePath)) {
-            snapshots.put(absolutePath, normalizedSnapshot);
-        }
-    }
-
-    public FileCollectionSnapshot build() {
-        if (snapshots.isEmpty()) {
-            return FileCollectionSnapshot.EMPTY;
-        }
-        return new DefaultFileCollectionSnapshot(snapshots, compareStrategy, pathNormalizationStrategy.isPathAbsolute());
-    }
+public interface FileCollectionSnapshotBuilder {
+    FileCollectionSnapshot build();
 }
